Moving infer_categorical_features to another file
